refactor: Extract logic to services
Introduced services, since I needed different implementations
for logic, such as logging in: you could either read them from
a file or wait for the user to input them. Also the main method
was getting a bit too large.

Now main serves as a composition root and uses the initialized
services to paste the contents from moodle to a source code file.